♻️ refactor(Categories): Admin 권한 확인 로직 위치를 Service -> Controller 레이어로 변경

+++ src/main/java/com/feelory/feelory_backend/words/controller/CategoriesController.java
@@ -2,6 +2,7 @@ package com.feelory.feelory_backend.words.controller;
 
 import com.feelory.feelory_backend.global.api.ApiResponse;
 import com.feelory.feelory_backend.global.api.SuccessCode;
+import com.feelory.feelory_backend.global.security.auth.jwt.JwtTokenProvider;
 import com.feelory.feelory_backend.words.model.*;
 import com.feelory.feelory_backend.words.service.CategoriesService;
 import io.swagger.v3.oas.annotations.Operation;
@@ -20,6 +21,7 @@ import org.springframework.web.bind.annotation.*;
 public class CategoriesController {
 
     private final CategoriesService categoriesService;
+    private final JwtTokenProvider jwtTokenProvider;
 
     @Operation(
             summary = "모든 카테고리 목록",
@@ -38,6 +40,8 @@ public class CategoriesController {
     )
     @PostMapping("")
     public ApiResponse<CategoryCreateResponse> postCategory(@Valid @RequestBody CategoryCreateRequest request) {
+        jwtTokenProvider.checkAdmin();
+
         CategoryCreateResponse response = categoriesService.registerCategory(request);
 
         return ApiResponse.success(response, SuccessCode.REGISTER_CATEGORY_SUCCESS);
@@ -49,6 +53,8 @@ public class CategoriesController {
     )
     @PatchMapping("")
     public ApiResponse<CategoryUpdateResponse> patchCategory(@Valid @RequestBody CategoryUpdateRequest request) {
+        jwtTokenProvider.checkAdmin();
+
         CategoryUpdateResponse response = categoriesService.modifyCategory(request);
 
         return ApiResponse.success(response, SuccessCode.UPDATE_CATEGORY_SUCCESS);
@@ -60,6 +66,8 @@ public class CategoriesController {
     )
     @DeleteMapping("")
     public ApiResponse<CategoryDeleteResponse> deleteCategory(@Valid CategoryDeleteRequest request) {
+        jwtTokenProvider.checkAdmin();
+
         CategoryDeleteResponse response = categoriesService.removeCategory(request);
 
         return ApiResponse.success(response, SuccessCode.DELETE_CATEGORY_SUCCESS);
